Implement IMapArea.contains(int, int) method
