+++ jsettlers.common/src/main/java/jsettlers/common/map/shapes/FreeMapArea.java
@@ -14,15 +14,14 @@
  *******************************************************************************/
 package jsettlers.common.map.shapes;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-
 import jsettlers.common.position.RelativePoint;
 import jsettlers.common.position.SRectangle;
 import jsettlers.common.position.ShortPoint2D;
 import jsettlers.common.utils.coordinates.CoordinateStream;
-import jsettlers.common.utils.coordinates.IBooleanCoordinateFunction;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 
 /**
  * This class gives a fast lookup (in O(1)) for contains if a MapArea is given by a list of n positions.<br>
@@ -234,16 +233,7 @@ public final class FreeMapArea implements IMapArea {
 
 	@Override
 	public CoordinateStream stream() {
-		return new CoordinateStream() {
-			@Override
-			public boolean iterate(IBooleanCoordinateFunction function) {
-				for (ShortPoint2D position : positions) {
-					if (!function.apply(position.x, position.y)) {
-						return false;
-					}
-				}
-				return true;
-			}
-		};
+		return CoordinateStream.fromList(positions);
 	}
+
 }
