+++ jsettlers.common/src/main/java/jsettlers/common/map/shapes/Parallelogram.java
@@ -14,12 +14,12 @@
  *******************************************************************************/
 package jsettlers.common.map.shapes;
 
-import java.util.Iterator;
-import java.util.NoSuchElementException;
-
 import jsettlers.common.position.ShortPoint2D;
 import jsettlers.common.utils.coordinates.CoordinateStream;
 import jsettlers.common.utils.coordinates.IBooleanCoordinateFunction;
+
+import java.util.Iterator;
+import java.util.NoSuchElementException;
 
 /**
  * This is a parallelogram on the map.
@@ -57,8 +57,11 @@ public class Parallelogram implements IMapArea {
 
 	@Override
 	public boolean contains(ShortPoint2D position) {
-		int x = position.x;
-		int y = position.y;
+		return contains(position.x, position.y);
+	}
+
+	@Override
+	public boolean contains(int x, int y) {
 		return x >= minx && x <= maxx && y >= miny && y <= maxy;
 	}
